// src/contract/enums.ts
export enum Role {
  Admin = "admin",
  DevOps = "devops",
  Developer = "developer",
}

export enum Priority {
  Low = "low",
  Medium = "medium",
  High = "high",
}

export enum Status {
  Todo = "todo",
  Doing = "doing",
  Done = "done",
}

export enum TaskPriority {
  Low = "low",
  Medium = "medium",
  High = "high",
}

export enum TaskStatus {
  Todo = "todo",
  Doing = "doing",
  Done = "done",
}